requirements.txt
appdirs==1.4.2
Django==1.10.5
funcsigs==1.0.2
h5py==2.6.0
image==1.5.5
Keras==1.2.2
mock==2.0.0
numpy==1.12.0
olefile==0.44
packaging==16.8
pbr==1.10.0
Pillow==4.0.0
protobuf==3.2.0
pyparsing==2.1.10
pyspark==2.1.0+hadoop2.7
PyYAML==3.12
scipy==0.18.1
six==1.10.0
tensorflow==1.0.0
Theano==0.8.2
